# backend/api-gateway/openapi/openapi.yaml
openapi: 3.1.0
info:
  title: LumenHorizon Public API
  version: 0.1.0
  description: API Gateway contract. Public read routes are anonymous; admin routes require JWT bearer auth.
servers:
  - url: https://api.lumenhorizon.com
    description: Public API endpoint
  - url: http://localhost:8080
    description: Local API Gateway development
security: []
...
paths:
  /api/v1/tiles/manifest:
    get:
      tags: [Tiles]
      summary: Get latest tile manifest
      operationId: getLatestTileManifest
      security: []
      responses:
        '200':
          $ref: '#/components/responses/TileManifest'
        '429':
          $ref: '#/components/responses/RateLimited'
        '500':
          $ref: '#/components/responses/InternalError'
        '503':
          $ref: '#/components/responses/ServiceUnavailable'
...
  /api/v1/tiles/sets:
    get:
      tags: [Tiles]
      summary: List tile sets
      operationId: listTileSets
      security: []
      parameters:
        - $ref: '#/components/parameters/Limit'
        - $ref: '#/components/parameters/Cursor'
      responses:
        '200':
          $ref: '#/components/responses/ObjectList'
        '400':
          $ref: '#/components/responses/InvalidRequest'
        '429':
          $ref: '#/components/responses/RateLimited'
        '503':
          $ref: '#/components/responses/ServiceUnavailable'
  /api/v1/tiles/classes:
    get:
      tags: [Tiles]
      summary: Get dark-sky class legend
      operationId: getTileClasses
      security: []
      responses:
        '200':
          $ref: '#/components/responses/ObjectResponse'
        '429':
          $ref: '#/components/responses/RateLimited'
...
  /api/v1/sites:
    get:
      tags: [Sites]
      summary: Search public observing sites
      operationId: searchSites
      security: []
      parameters:
        - name: lat
          in: query
          schema:
            type: number
            minimum: -90
            maximum: 90
        - name: lon
          in: query
          schema:
            type: number
            minimum: -180
            maximum: 180
        - name: radius_km
          in: query
          schema:
            type: number
            minimum: 1
            maximum: 250
            default: 50
        - name: west
          in: query
          schema:
            type: number
            minimum: -180
            maximum: 180
        - name: south
          in: query
          schema:
            type: number
            minimum: -90
            maximum: 90
        - name: east
          in: query
          schema:
            type: number
            minimum: -180
            maximum: 180
        - name: north
          in: query
          schema:
            type: number
            minimum: -90
            maximum: 90
        - $ref: '#/components/parameters/Limit'
        - $ref: '#/components/parameters/Cursor'
      responses:
        '200':
          $ref: '#/components/responses/SiteList'
        '400':
          $ref: '#/components/responses/InvalidRequest'
        '429':
          $ref: '#/components/responses/RateLimited'
        '503':
          $ref: '#/components/responses/ServiceUnavailable'
...
components:
  securitySchemes:
    AdminJwt:
      type: http
      scheme: bearer
      bearerFormat: JWT
      description: RS256 JWT validated against configured issuer, audience, JWKS URL, required subject, expiry, and admin role claim.
  parameters:
...
    Limit:
      name: limit
      in: query
      schema:
        type: integer
        minimum: 1
        maximum: 100
        default: 50
    Cursor:
      name: cursor
      in: query
      schema:
        type: string
        minLength: 1
        maxLength: 2048
  responses:
    TileManifest:
      description: Tile manifest envelope
      content:
        application/json:
          schema:
            allOf:
              - $ref: '#/components/schemas/SuccessEnvelope'
              - type: object
                properties:
                  data:
                    $ref: '#/components/schemas/TileManifest'
    SiteList:
      description: Site list envelope
      content:
        application/json:
          schema:
            allOf:
              - $ref: '#/components/schemas/SuccessEnvelope'
              - type: object
                properties:
                  data:
                    type: array
                    items:
                      $ref: '#/components/schemas/Site'
...
    ObjectResponse:
      description: Object envelope
      content:
        application/json:
          schema:
            $ref: '#/components/schemas/SuccessEnvelope'
    ObjectList:
      description: List envelope
      content:
        application/json:
          schema:
            $ref: '#/components/schemas/SuccessEnvelope'
...
    InvalidRequest:
      description: Invalid request
      content:
        application/json:
          schema:
            $ref: '#/components/schemas/ErrorEnvelope'
...
    RateLimited:
      description: Rate limit exceeded
      headers:
        Retry-After:
          schema:
            type: integer
            minimum: 1
      content:
        application/json:
          schema:
            $ref: '#/components/schemas/ErrorEnvelope'
    InternalError:
      description: Unexpected server failure
      content:
        application/json:
          schema:
            $ref: '#/components/schemas/ErrorEnvelope'
...
    ServiceUnavailable:
      description: Required dependency unavailable or service not ready
      content:
        application/json:
          schema:
            $ref: '#/components/schemas/ErrorEnvelope'
...
  schemas:
    SuccessEnvelope:
      type: object
      required: [data, meta, error]
      additionalProperties: false
      properties:
        data: true
        meta:
          $ref: '#/components/schemas/EnvelopeMeta'
        error:
          type: 'null'
    ErrorEnvelope:
      type: object
      required: [data, meta, error]
      additionalProperties: false
      properties:
        data:
          type: 'null'
        meta:
          $ref: '#/components/schemas/EnvelopeMeta'
        error:
          $ref: '#/components/schemas/ErrorBody'
    EnvelopeMeta:
      type: object
      required: [request_id, timestamp]
      additionalProperties: false
      properties:
        request_id:
          type: string
          format: uuid
        timestamp:
          type: string
          format: date-time
        next_cursor:
          type: string
          nullable: true
    ErrorBody:
      type: object
      required: [code, message]
      additionalProperties: false
      properties:
        code:
          type: string
          enum:
            - invalid_request
            - unauthenticated
            - forbidden
            - not_found
            - tile_not_found
            - conflict
            - unprocessable_entity
            - rate_limited
            - internal_error
            - upstream_error
            - service_unavailable
            - tile_unavailable
        message:
          type: string
        details:
          type: object
          additionalProperties: true
          nullable: true
    TileManifest:
      type: object
      required: [tile_set_id, dataset_date, generated_at, classification_version, render_version, format, tile_size, min_zoom, max_native_zoom, max_display_zoom, bounds, tile_url_template]
      additionalProperties: true
      properties:
        tile_set_id:
          type: string
        dataset_date:
          type: string
          format: date
        generated_at:
          type: string
          format: date-time
        classification_version:
          type: string
        render_version:
          type: string
        format:
          type: string
          enum: [png]
        tile_size:
          type: integer
          enum: [256]
        min_zoom:
          type: integer
        max_native_zoom:
          type: integer
        max_display_zoom:
          type: integer
        bounds:
          $ref: '#/components/schemas/Bounds'
        tile_url_template:
          type: string
    Bounds:
      type: object
      required: [west, south, east, north]
      additionalProperties: false
      properties:
        west:
          type: number
          minimum: -180
          maximum: 180
        south:
          type: number
          minimum: -90
          maximum: 90
        east:
          type: number
          minimum: -180
          maximum: 180
        north:
          type: number
          minimum: -90
          maximum: 90
    Site:
      type: object
      required: [site_id, name, latitude, longitude]
      additionalProperties: true
      properties:
        site_id:
          type: string
          format: uuid
        name:
          type: string
        latitude:
          type: number
          minimum: -90
          maximum: 90
        longitude:
          type: number
          minimum: -180
          maximum: 180
...